Update jquery.selectionBox.js
Added some code to use the id of an input as the class, so that individual inputs could be specifically styled. Added z-index adjustments to help make sure input options show above other items on a page (z-index of 99), then returns to z-index of 1 after being done with. Made open options drop downs hide when opening a different options list.